+++ NEWS.md
@@ -3,7 +3,8 @@
 
 * Internal fixes in prevision of using `tidyselect` within `dplyr`.
 
-* `vars_select()` now supports unquoting of named character vectors.
+* `vars_select()` and `vars_rename()` now correctly support unquoting
+  character vectors that have names.
 
 
 # tidyselect 0.2.2